Add support for FTP
Supporting FTP is an odd choice. Most browsers don’t support it. I only
found two browsers, Konqueror [1] and Pale Moon [2], that still support
FTP, and one of them (Konqueror) doesn’t even support loading Web pages
over FTP (it downloads them instead).

One of the goals that I have for this site is protocol diversity. I want
people to be able to choose the protocol that they prefer. If someone
wants to start using BSD [3] or Linux [4] instead of Windows [5], then
they should do so. The competition benefits everyone. If someone wants
to use IPFS [6] or Hypercore [7] instead of HTTP [8], then they should
do so. The competition benefits everyone.

That being said, I highly doubt that anyone thinks that FTP is superior
to HTTPS. The main goal of adding FTP support is to make sure that I can
add support for more protocols in the future. I was originally planning
on adding IPFS support for this purpose, but that proved to be
too complicated.

[1]: <https://apps.kde.org/konqueror/>
[2]: <https://www.palemoon.org/>
[3]: <https://www.geeklan.co.uk/?p=2061>
[4]: <https://kernel.org/linux.html>
[5]: <https://www.microsoft.com/windows/>
[6]: <https://ipfs.io/>
[7]: <https://hypercore-protocol.org/>
[8]: <https://httpwg.org/>