Stop using the HTML <base> element
Originally, I wanted to use <base> to help me make the site available
under multiple URI schemes. Unfortunately, <base> can’t really be used
on IPFS. In order to use <base>, you have to know a site’s URL before
you finish writing it. Since IPFS is content addressed, you can’t know
what a page’s URL will be until after your done writing it. In other
words, adding a base URL to an IPFS page would change the page’s URL,
which means that you have to change the base URL, which means that the
page’s URL changes, which means that you have to change the base URL
again, which means that the page’s URL changes, which means that you
have to change the base URL again, which means…

In addition, removing support for base URLs shrunk the build tool and
broke the Neocities redirect version of the site. I deleted the
Neocities redirect version of the site. It was always supposed to be a
version of the site that was frozen in time, so it’s OK if it doesn’t
get updated. I want to direct people to the jasonyundt.website instead,
anyway.