Create blog post about accepting patches
This change makes main.css fail to validate. It actually is valid, but
the version of v.Nu that I’m using isn’t new enough. The best way to fix
that problem would be to update html5validator
(<https://github.com/svenkreiss/html5validator>) to use a newer version
of v.Nu. The problem is that it’s unclear which version of v.Nu
html5validator should use (see
<https://github.com/validator/validator/issues/1354>).