Remove version numbers from most prerequsites
At the moment, it’s not feasible to test building the site with
different versions of all of the prerequisites. As a result, I won’t
really know if a specific version of, say, env will work. I just know
that whatever version of env that’s currently on my system works.