Make nav bar appear on homepage
I also made it so that the nav bar on the homepage doesn’t include a
link to the homepage because the home link implies that you’re going
back to the homepage.

At the moment, the nav bar is kind of pointless on the homepage, but it
will be more useful once I add more links to it.