Specify legal terms that apply to commit metadata
This change aligns commit metadata with FOSS expectations. While FOSS
users and contributors expect software to potentially be copyrighted,
they generally don’t expect commit messages to be copyrighted. Most Git
users are willing to push commits wherever, not thinking about the fact
that they might be infringing on the copyright of the commit messages.