Use Python 3.x instead of Python 3.8 in particular
python38 was deleted from Nixpkgs [1] so I can no longer use it. I
switch from an attribute specifying a specific minor version of Python
to an attribute that only specifies a specific major version of Python.
Hopefully, this will mean that I don’t have to keep changing the
attribute in the future.

[1]: <https://github.com/NixOS/nixpkgs/pull/285779#event-12021105668>